Enabled window resizing for desktop
This change enables window resizing for desktop applications for D3D12
and Vulkan. Each API takes a different path for resizing due how their
swapchains are handled.

D3D12: IDXGISwapChain::ResizeBuffers is used. grfx::Swapchain::Resize()
was added to handle window resize events. There is a weird workaround
for D3D12 to correctly invalidate the client area on resizes.

Vulkan: The swapchain is destroyed and recreated.
grfx::Swapchain::Resize() returns an error on Vulkan.

Changes:
- Removed the surface caps caching for Vulkan. The caching was causing
  the surface width/height to be locked into when the surface was
  created.
- Fixed AddRef/Releaase related bug in D3D12's image class that was
  causing resource references not to be released when the resource was
  external.
- Enabled window resizing for 01_triangle sample.
- Fixed some casting warnings

+++ src/ppx/grfx/dx12/dx12_descriptor.cpp
@@ -336,7 +336,7 @@ Result DescriptorSet::UpdateDescriptors(uint32_t writeCount, const grfx::WriteDe
                 desc.Format                           = DXGI_FORMAT_R32_TYPELESS;
                 desc.ViewDimension                    = D3D12_UAV_DIMENSION_BUFFER;
                 desc.Buffer.FirstElement              = srcWrite.bufferOffset / 4;
-                desc.Buffer.NumElements               = sizeInBytes / 4;
+                desc.Buffer.NumElements               = static_cast<UINT>(sizeInBytes / 4);
                 desc.Buffer.StructureByteStride       = 0;
                 desc.Buffer.Flags                     = D3D12_BUFFER_UAV_FLAG_RAW;
